requirements.txt
msal-streamlit-authentication==1.0.9
azure-ai-formrecognizer
streamlit==1.32.2
azure-search-documents==11.2.0
scikit-learn==0.24.2
streamlit-lottie>=0.0.5
azure-core
azure-identity
requests==2.28.1
urllib3==1.26.12
python-dotenv==1.0.1
pyodbc==5.1.0
